app/requirements.txt
numpy
pandas
pillow
streamlit
onnxruntime
scikit-learn
